+++ src/app/app.module.ts
@@ -71,11 +71,9 @@ import { PermissionsManagerComponent } from './components/permission-manager/per
 import { AppRouteReuseStrategy } from './app.routes.strategy';
 import { AppInfoDrawerModule } from './components/info-drawer/info.drawer.module';
 import { DirectivesModule } from './directives/directives.module';
-import { ToggleInfoDrawerComponent } from './components/toolbar/toggle-info-drawer/toggle-info-drawer.component';
-import { DocumentDisplayModeComponent } from './components/toolbar/document-display-mode/document-display-mode.component';
-import { ToggleFavoriteComponent } from './components/toolbar/toggle-favorite/toggle-favorite.component';
 import { ContextMenuModule } from './components/context-menu/context-menu.module';
 import { ExtensionsModule } from '@alfresco/adf-extensions';
+import { AppToolbarModule } from './components/toolbar/toolbar.module';
 
 @NgModule({
     imports: [
@@ -97,7 +95,8 @@ import { ExtensionsModule } from '@alfresco/adf-extensions';
 
         DirectivesModule,
         ContextMenuModule.forRoot(),
-        AppInfoDrawerModule
+        AppInfoDrawerModule,
+        AppToolbarModule
     ],
     declarations: [
         AppComponent,
@@ -123,10 +122,7 @@ import { ExtensionsModule } from '@alfresco/adf-extensions';
         PermissionsManagerComponent,
         SearchResultsComponent,
         SettingsComponent,
-        SharedLinkViewComponent,
-        ToggleInfoDrawerComponent,
-        DocumentDisplayModeComponent,
-        ToggleFavoriteComponent
+        SharedLinkViewComponent
     ],
     providers: [
         { provide: RouteReuseStrategy, useClass: AppRouteReuseStrategy },
@@ -149,10 +145,7 @@ import { ExtensionsModule } from '@alfresco/adf-extensions';
     entryComponents: [
         LibraryDialogComponent,
         NodeVersionsDialogComponent,
-        NodePermissionsDialogComponent,
-        ToggleInfoDrawerComponent,
-        DocumentDisplayModeComponent,
-        ToggleFavoriteComponent
+        NodePermissionsDialogComponent
     ],
     bootstrap: [AppComponent]
 })
