+++ src/app/components/toolbar/toolbar-action/toolbar-action.component.ts
@@ -1,0 +1,56 @@
+/*!
+ * @license
+ * Alfresco Example Content Application
+ *
+ * Copyright (C) 2005 - 2018 Alfresco Software Limited
+ *
+ * This file is part of the Alfresco Example Content Application.
+ * If the software was purchased under a paid Alfresco license, the terms of
+ * the paid license agreement will prevail.  Otherwise, the software is
+ * provided under the following open source license terms:
+ *
+ * The Alfresco Example Content Application is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * The Alfresco Example Content Application is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with Alfresco. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import {
+    Component,
+    ViewEncapsulation,
+    ChangeDetectionStrategy,
+    Input
+} from '@angular/core';
+import { AppStore } from '../../../store/states';
+import { Store } from '@ngrx/store';
+import { ContentActionRef } from '@alfresco/adf-extensions';
+import { AppExtensionService } from '../../../extensions/extension.service';
+
+@Component({
+    selector: 'aca-toolbar-action',
+    templateUrl: './toolbar-action.component.html',
+    encapsulation: ViewEncapsulation.None,
+    changeDetection: ChangeDetectionStrategy.OnPush,
+    host: { class: 'aca-toolbar-action' }
+})
+export class ToolbarActionComponent {
+    @Input() type = 'icon-button';
+    @Input() entry: ContentActionRef;
+
+    constructor(
+        protected store: Store<AppStore>,
+        protected extensions: AppExtensionService
+    ) {}
+
+    trackByActionId(index: number, action: ContentActionRef) {
+        return action.id;
+    }
+}
