+++ src/app/components/toolbar/toolbar-button/toolbar-button.component.html
@@ -1,0 +1,28 @@
+<ng-container [ngSwitch]="type">
+    <ng-container *ngSwitchCase="'icon-button'">
+        <button
+            [id]="actionRef.id"
+            mat-icon-button
+            color="primary"
+            [attr.title]="(actionRef.description || actionRef.title) | translate"
+            (click)="runAction()">
+            <mat-icon>{{ actionRef.icon }}</mat-icon>
+        </button>
+    </ng-container>
+    <ng-container *ngSwitchCase="'menu-item'">
+        <button
+            [id]="actionRef.id"
+            mat-menu-item
+            color="primary"
+            [disabled]="actionRef.disabled"
+            [attr.title]="(
+                actionRef.disabled
+                ? actionRef['description-disabled']
+                : actionRef.description || actionRef.title
+            ) | translate"
+            (click)="runAction()">
+            <mat-icon>{{ actionRef.icon }}</mat-icon>
+            <span>{{ actionRef.title | translate }}</span>
+        </button>
+    </ng-container>
+</ng-container>
